requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
python-multipart